LogUI hot fix, introduced at mew message optimisation, that break Application Log POrt compile

+++ java/org/contikios/cooja/dialogs/LogUI.java
@@ -111,7 +111,7 @@ public abstract class LogUI extends Log {
       }
   }
 
-  public void dataReceived(byte data) {
+  public void dataReceived(int data) {
     if (data == '\n') {
       /* Notify observers of new log */
       lastLogMessage = newMessage.toString();
